lesson-03-production-data-pipelines/invoice-extraction/requirements.txt
pdfplumber>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
